+++ c++/input.h
@@ -2,4 +2,4 @@
 #include "solution.h"
 #include <tuple>
 
-std::tuple<Edgelist, int64_t, int64_t, int64_t> parse_input(std::string filename);
+std::tuple<Edgelist, __int128_t, __int128_t, __int128_t> parse_input(std::string filename);
